requirements.txt
click==8.0.3
requests==2.27.1
msgpack==1.0.3
python-dotenv==0.19.2
